requirements.txt
Flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
python-engineio>=4.9.0
eventlet>=0.33.3
ccxt>=4.2.0
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.11.0
